+++ Source/WebCore/svg/SVGElement.cpp
@@ -26,6 +26,8 @@
 #include "config.h"
 #include "SVGElement.h"
 
+#include "CSSComputedStyleDeclaration.h"
+#include "CSSPrimitiveValueMappings.h"
 #include "CSSPropertyParser.h"
 #include "Document.h"
 #include "ElementChildIterator.h"
@@ -710,6 +712,20 @@ const RenderStyle* SVGElement::computedStyle(PseudoId pseudoElementSpecifier)
     }
 
     return m_svgRareData->overrideComputedStyle(*this, parentStyle);
+}
+
+ColorInterpolation SVGElement::colorInterpolation() const
+{
+    if (auto renderer = this->renderer())
+        return renderer->style().svgStyle().colorInterpolationFilters();
+
+    // Try to determine the property value from the computed style.
+    if (auto value = ComputedStyleExtractor(const_cast<SVGElement*>(this)).propertyValue(CSSPropertyColorInterpolationFilters, DoNotUpdateLayout)) {
+        if (is<CSSPrimitiveValue>(value))
+            return downcast<CSSPrimitiveValue>(*value);
+    }
+
+    return ColorInterpolation::Auto;
 }
 
 QualifiedName SVGElement::animatableAttributeForName(const AtomString& localName)
