+++ Source/WebCore/svg/graphics/filters/SVGFilter.h
@@ -39,6 +39,8 @@ public:
     static RefPtr<SVGFilter> create(SVGFilterElement&, RenderingMode, const FloatSize& filterScale, ClipOperation, const FloatRect& filterRegion, const FloatRect& targetBoundingBox, const GraphicsContext& destinationContext);
     WEBCORE_EXPORT static RefPtr<SVGFilter> create(const FloatRect& targetBoundingBox, SVGUnitTypes::SVGUnitType primitiveUnits, SVGFilterExpression&&);
 
+    static IntOutsets calculateOutsets(SVGFilterElement&, const FloatRect& targetBoundingBox);
+
     FloatRect targetBoundingBox() const { return m_targetBoundingBox; }
     SVGUnitTypes::SVGUnitType primitiveUnits() const { return m_primitiveUnits; }
 
@@ -56,6 +58,7 @@ private:
     SVGFilter(RenderingMode, const FloatSize& filterScale, ClipOperation, const FloatRect& filterRegion, const FloatRect& targetBoundingBox, SVGUnitTypes::SVGUnitType primitiveUnits);
     SVGFilter(const FloatRect& targetBoundingBox, SVGUnitTypes::SVGUnitType primitiveUnits, SVGFilterExpression&&);
 
+    static std::optional<SVGFilterExpression> buildExpression(SVGFilterElement&, const SVGFilter&, const GraphicsContext& destinationContext);
     void setExpression(SVGFilterExpression&& expression) { m_expression = WTFMove(expression); }
 
     FloatSize resolvedSize(const FloatSize&) const final;
